add co-rtdetr

+++ mmdet/models/dense_heads/rpn_head.py
@@ -281,8 +281,9 @@ class RPNHead(AnchorHead):
 
         if results.bboxes.numel() > 0:
             bboxes = get_box_tensor(results.bboxes)
-            det_bboxes, keep_idxs = batched_nms(bboxes, results.scores,
-                                                results.level_ids, cfg.nms)
+            det_bboxes, keep_idxs = batched_nms(
+                bboxes, results.scores.type_as(bboxes),
+                results.level_ids, cfg.nms)
             results = results[keep_idxs]
             # some nms would reweight the score, such as softnms
             results.scores = det_bboxes[:, -1]
